+++ CHANGELOG.md
@@ -8,8 +8,9 @@
 
  - Fix inability to activate Edge Conditionals ([#381](https://github.com/kizniche/mycodo/issues/381))
  - Fix inability to add new gauges or graphs to the dashboard ([#384](https://github.com/kizniche/mycodo/issues/384))
- - Fix issue with UART Atlas pH initialization ([#382](https://github.com/kizniche/mycodo/issues/382))
+ - Fix issues with UART Atlas pH Input device ([#382](https://github.com/kizniche/mycodo/issues/382))
  - Fix issue with Atlas pH calibration ([#382](https://github.com/kizniche/mycodo/issues/382))
+ - Fix issue with caching of Camera images on the Dashboard
 
 
 ## 5.5.9 (2018-01-14)